Move ostream, remove some extra headers
