+++ include/boost/histogram/python/accumulators/ostream.hpp
