+++ include/boost/histogram/python/accumulators/weighted_mean.hpp
@@ -12,10 +12,8 @@
 
 #pragma once
 
-#include <boost/assert.hpp>
 #include <boost/core/nvp.hpp>
 #include <boost/histogram/weight.hpp>
-#include <type_traits>
 
 namespace accumulators {
