+++ include/boost/histogram/python/register_accumulator.hpp
@@ -7,7 +7,7 @@
 
 #include <boost/histogram/python/pybind11.hpp>
 
-#include <boost/histogram/python/accumulators_ostream.hpp>
+#include <boost/histogram/python/accumulators/ostream.hpp>
 #include <boost/histogram/python/make_pickle.hpp>
 
 #include <pybind11/operators.h>
